Issue #35 - Menu em Cascata
Implementação inicial do menu em cascata #35
Pela falta dos outros botões utilizei dois objetos aleatorios como
~botões~
Generalizei alguns scripts para animação de botões.
