+++ DOCES - Unity/Assets/Camera/DropdownMenu.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e7ad31e77c246644cb32434b2cd98088
+folderAsset: yes
+timeCreated: 1475519757
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
