+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8ebb6a71e735ac54fa2a782135cd3fed
+folderAsset: yes
+timeCreated: 1475522863
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
