+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropDownDeactivating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e0ae64d30e2a3e4fab7f34d6587cf8e
+timeCreated: 1475540374
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
