+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownActivated.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0926b22da6a3e27469e4449f33dae664
+timeCreated: 1475538836
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
