+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownActivating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c4c9e2f28a6ed845acdb4f0ff3ec13e
+timeCreated: 1475540358
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
