+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownDeactivated.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c21530468969b34ea8b59965f41be06
+timeCreated: 1475538943
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
