+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownDeselected.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fcab3b42a1085a4791483ee422ef716
+timeCreated: 1475536319
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
