+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownMenuAnimator.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 34797df0acb5f3a49932836b60764da6
+timeCreated: 1475536085
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
