+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownNormal.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a4a21f4434aeaa5439135835524c034d
+timeCreated: 1475536307
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
