+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownPressed.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39062bb83b9a0984abcc1d06e14509be
+timeCreated: 1475536338
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
