+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/DropdownSelected.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 78d9bfda5e38af94581ddf7b9013e9f6
+timeCreated: 1475536308
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
