+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/FirstButton.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 2fed0884f55c291418038afb86f0843f
+folderAsset: yes
+timeCreated: 1475541836
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
