+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/FirstButton/New Sprite.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b97b215a3b690064f859fcdecd2d607f
+timeCreated: 1475541639
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
