+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/FirstButton/b1Activating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bcb1f7a6facd68640b8d3c0c4ccc8266
+timeCreated: 1475541639
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
