+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/FirstButton/b1Deactivating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa9e15de24ff9ba4eb90825a79fc5648
+timeCreated: 1475541659
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
