+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/SecondButton.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 3cefa9d80e3e2f94b80599dd4ff345bd
+folderAsset: yes
+timeCreated: 1475542142
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
