+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/SecondButton/New Sprite (1).controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d199a3449cc04f4c861f50a9c5d5155
+timeCreated: 1475542161
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
