+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/SecondButton/b2Activating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d036f63d2567df4bb8adf483e5a4588
+timeCreated: 1475542161
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
