+++ DOCES - Unity/Assets/Camera/DropdownMenu/Animations/SecondButton/b2Deactivating.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c1d56984b4a3f7341a095142c15bd965
+timeCreated: 1475542268
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
