+++ DOCES - Unity/Assets/Camera/DropdownMenu/Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 51a4761503b29d543ad46e252c94a771
+folderAsset: yes
+timeCreated: 1475519781
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
