+++ DOCES - Unity/Assets/Camera/DropdownMenu/Scripts/DropdownMenu.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d02f97a0def990d4ea312e0e73a329fc
+timeCreated: 1475519795
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
