+++ DOCES - Unity/Assets/Camera/DropdownMenu/Sprite.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: eb178f9e6a8dfa44aad30790350ad023
+folderAsset: yes
+timeCreated: 1475521048
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
