+++ DOCES - Unity/Assets/GenericButton.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e7893b0944e263f4b8902629a6d692b5
+folderAsset: yes
+timeCreated: 1475524353
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
