+++ DOCES - Unity/Assets/GenericButton/GenericButtonAnimation.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f22bd53d3422be44dbe4b264b7a58d06
+folderAsset: yes
+timeCreated: 1475523181
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
