+++ DOCES - Unity/Assets/GenericButton/GenericButtonAnimation/GenericButtonAnimatorOverride.overrideController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aac042cbbd7fc7244a717ae3643bbe66
+timeCreated: 1475523198
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
