+++ DOCES - Unity/Assets/StartScreen/New Game Button/Animation/NewGameButtonAnimator.controller.meta
